이클립스 상 오류 해결 및 파일 정리

+++ src/summary/seungha/Ch03/Basic/OperatorEx6.java
@@ -4,7 +4,8 @@ class OperatorEx6 {
 	public static void main(String[] args) {
 		byte a = 10;
 		byte b = 20;
-		byte c = a + b;		// 컴파일 에러 발생. 명시적 형변환 필요
+//		byte c = a + b;		// 컴파일 에러 발생. 명시적 형변환 필요
+		byte c = (byte)(a+b);
 		System.out.println(c);
 	}
 }
